Fix memory leak caused by Iterable/Iterator mixups (#974)
Major change here is to strictly enforce that compute(), map(), and
related functionality returns an Iterator, not an Iterable, since the
former is more likely to be O(1) memory (misuses of tee aside...),
whereas the latter is more likely to be O(N) memory.

Still haven't figured out the duckdb memory leak; seems like something
in our mess of join views and nested views is causing duckdb to pull the
entire parquet file into memory, even if the final query is only
requesting row_id and one other column. The ORDER BY ROWID is likely
contributing to the issue, but removing the ORDER BY does not fully fix
the "pull it all into memory" issue.