`dataset.cluster()` now uses `transform()` which uses `map()` (#1002)
Make `dataset.cluster()` use `transform()` which uses `map()` for the
entire pipeline.

This completes the refactor needed so that signals can call `map()`
instead.

- `dataset.cluster()` now writes the output to a sibling column (next to
the original path, not under it)
- `dataset.map()` can now take optional schema. This is needed so that
we can tag `cluster_id `as categorical. And it will unlock signals to
call `map()`.
- `dataset.map()` can now take `embedding:str`. When provided, the map
function will be called with the embedding spans for that item. This is
needed for any computation that needs the embeddings (e.g. clustering).
And it will unlock any `VectorSignal`s to call `map()` inside
`compute_vector()`.
- `combine_columns` is now gone from `map()`. It will always combine,
which is simpler since we will always write to a sibling
- Add utils for determining the sibling path and the common ancestor
since we sometimes need our `map()` to see both the original info and
the sibling info at once.